🐛🚨 fix translation indicator and linter errors
