+++ rogue-thi-app/components/modal/DashboardModal.js
@@ -1,10 +1,10 @@
+import { DashboardContext, ShowDashboardModal } from '../../pages/_app'
 import { faChevronDown, faChevronUp, faTrash, faTrashRestore } from '@fortawesome/free-solid-svg-icons'
 import { useContext, useRef } from 'react'
 import Button from 'react-bootstrap/Button'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import ListGroup from 'react-bootstrap/ListGroup'
 import Modal from 'react-bootstrap/Modal'
-import { DashboardContext, ShowDashboardModal } from '../../pages/_app'
 import styles from '../../styles/Personalize.module.css'
 
 /**
